2 Fixes for Debian Bug#623425
Both patches are by Jonathan Nieder <jrnieder@gmail.com>

Ever since psmisc 22.4 (2007-04-08), "killall <path>" compares
the target of the /proc/<pid>/exe link to <path> when comparison
of inode numbers fails.  But the code to do this has two problems:

 - readlink does not NUL-terminate, but we use strcmp to
   compare strings.  Probably this hasn't been a problem so far
   because (1) the on-stack buffer happened to have zeroes in the
   right places or (2) some implementations of readlink might
   happen to NUL-terminate their result when convenient anyway.

 - it relies on PATH_MAX to determine the size of the buffer,
   so the code fails to build from source on platforms (like the
   Hurd) that have no global PATH_MAX.

Fix both by using a buffer of size strlen("<path>") + 1 and comparing
the link target to <path> with memcmp after checking that it fit in
the buffer.

For consistency with the surrounding code, the pid is considered not
to match if malloc or readlink fails.

---------------
The Hurd has most of the expected /proc/<pid> hierarchy but no
/proc/self.  Without this change, running "sleep 10 & killall sleep"
on that platform produces the confusing message:
    /proc is empty (not mounted ?)